Acute Stroke Patient and Non Acute patient
The Images Data sets contain a collection of pictures of individuals who have been diagnosed with acute stroke and individuals who have not been diagnosed with acute stroke. The dataset includes pictures of the face of each individual, which can be used to identify different characteristics and features that may be indicative of a stroke. These images can be used to train Deep learning models to detect and diagnose stroke in patients. The dataset is a valuable resource for researchers and healthcare professionals working in the field of stroke medicine.