Source Code File
The Deep Learning Model used in this dataset utilizes data augmentation techniques on images to improve the overall accuracy of the model. The model has a learning rate of 0.1 and is trained for 10 epochs. The model also uses k-fold cross-validation with a k-value of 3 to improve the robustness of the model and prevent overfitting. After training, the model achieved an accuracy of 94%. This high accuracy indicates that the model is able to accurately detect and diagnose acute stroke in patients using facial images. A model is a valuable tool for healthcare professionals and researchers working in the field of stroke medicine.